requirements.txt
pdfminer.six==20221105
PyPDF2==3.0.1
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0 